refactor: rename systax builder module
