+++ pyrdfj2/__init__.py
@@ -1,5 +1,5 @@
 from pyrdfj2.j2_functions import Filters, Functions
 from pyrdfj2.pyrdfj2 import RDFSyntaxBuilder
-from pyrdfj2.rdf_syntax_builder import J2RDFSyntaxBuilder
+from pyrdfj2.syntax_builder import J2RDFSyntaxBuilder
 
 __all__ = ["Filters", "Functions", "RDFSyntaxBuilder", "J2RDFSyntaxBuilder"]
